refactor: splide removing (part 1)
