+++ package.json
@@ -27,7 +27,6 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.5",
-    "@splidejs/react-splide": "^0.7.12",
     "@vanilla-extract/css": "^1.20.0",
     "embla-carousel-autoplay": "^8.6.0",
     "embla-carousel-react": "^8.6.0",
