+++ src/Template/BarSlider/BarSlider.css.ts
@@ -1,12 +1,11 @@
 import { style } from '@vanilla-extract/css';
 import { breakpoints } from '@/styles/theme.css';
 
-export const barSlider = style({
+export const viewport = style({
   borderRadius: '10px',
   border: 'solid 1px var(--secondaryClr)',
   margin: '1rem 0',
   overflow: 'hidden',
-  position: 'relative',
   cursor: 'grab',
   width: '100%',
   aspectRatio: '16/9',
@@ -15,15 +14,22 @@ export const barSlider = style({
   },
 });
 
-export const imgContainer16x9 = style({
+export const track = style({
+  display: 'flex',
+  height: '100%',
+});
+
+export const slide = style({
+  flex: '0 0 100%',
+  minWidth: 0,
   position: 'relative',
-  width: '100%',
   aspectRatio: '16/9',
   overflow: 'hidden',
 });
 
-export const imgContainerImg = style({
+export const img = style({
   borderRadius: '10px',
   width: '100%',
   height: '100%',
+  objectFit: 'cover',
 });
