+++ src/Template/BarSlider/BarSlider.tsx
@@ -1,72 +1,32 @@
 import { memo } from 'react';
-import { SUSHI_BAR_URL } from '@/utils/constants';
-import { Splide, SplideSlide } from '@splidejs/react-splide';
+import Autoplay from 'embla-carousel-autoplay';
+import useEmblaCarousel from 'embla-carousel-react';
 import * as s from './BarSlider.css';
 
-// const BarSliderSlide = ({ slideKey }: {
-//   slideKey: string;
-// }) => {
-//   return (
-//     <SplideSlide key={slideKey}>
-//       <div className="imgContainer_16na9">
-//         <img
-//           srcSet={`Images/kombo/${slideKey}_sm.webp 500w,
-//                    Images/kombo/${slideKey}_md.webp 680w,
-//                    Images/kombo/${slideKey}_lg.webp 1920w`}
-//           alt={slideKey}
-//           className="imgContainer__img transition"
-//           src={`Images/kombo/${slideKey}_lg.webp`}
-//         />
-//       </div>
-//     </SplideSlide>
-//   );
-// };
+const slideKeys = ['kombo1', 'kombo2'];
 
-const DeliverySlide = ({ slideKey }: {
-  slideKey: string;
-}) => {
-  return (
-    <SplideSlide key={slideKey}>
-      <div className={s.imgContainer16x9}>
-        <img
-          srcSet={`Images/kombo/${slideKey}_sm.webp 500w,
-                   Images/kombo/${slideKey}_md.webp 680w,
-                   Images/kombo/${slideKey}_lg.webp 1920w`}
-          alt={slideKey}
-          className={`swSlide__img ${s.imgContainerImg}`}
-          src={`Images/kombo/${slideKey}_lg.webp`}
-        />
-      </div>
-    </SplideSlide>
+export const BarSlider = memo(function BarSlider() {
+  const [emblaRef] = useEmblaCarousel(
+    { loop: true, direction: 'ltr' },
+    [Autoplay({ delay: 3500, stopOnMouseEnter: true, stopOnInteraction: false })],
   );
-};
 
-const slideKeys = ['sushi_delivery'];
-
-const BarSlider = memo(function BarSlider() {
   return (
-    <Splide
-      options={{
-        perPage: 1,
-        perMove: 1,
-        pagination: false,
-        arrows: false,
-        // delivery -----------
-        drag: false,
-        // bar ----------------
-        // type: 'loop',
-        // direction: 'rtl',
-        // autoplay: true,
-        // interval: 3500,
-        // pauseOnHover: true,
-      }}
-      className={s.barSlider}
-    >
-      {slideKeys.map(key => (
-        <DeliverySlide key={key} slideKey={key} />
-      ))}
-    </Splide>
+    <div className={s.viewport} ref={emblaRef}>
+      <div className={s.track}>
+        {slideKeys.map(key => (
+          <div className={s.slide} key={key}>
+            <img
+              srcSet={`Images/kombo/${key}_sm.webp 500w,
+                       Images/kombo/${key}_md.webp 680w,
+                       Images/kombo/${key}_lg.webp 1920w`}
+              alt={key}
+              className={`swSlide__img ${s.img}`}
+              src={`Images/kombo/${key}_lg.webp`}
+            />
+          </div>
+        ))}
+      </div>
+    </div>
   );
 });
-
-export default BarSlider;
