+++ src/Template/FilmsSlider/FilmsSlider.tsx
@@ -1,16 +1,24 @@
-import { memo, useMemo } from 'react';
 import { FilmImg } from '@/Template/FilmImg';
 import { Link, useMatch } from 'react-router-dom';
+import { memo, useMemo, useCallback } from 'react';
+import useEmblaCarousel from 'embla-carousel-react';
 import { PRE_SHOW_SERVICE } from '@/utils/constants';
-import { Splide, SplideSlide } from '@splidejs/react-splide';
 import { removeLineBreaks } from '@/utils/formatTextWithLineBreaks';
-import { useMobilePaddings } from '../useMobilePaddings';
 import type { CinemaStateT } from '../../REDUX/cinema/cinemaReducerT';
 import * as s from './FilmsSlider.css';
 
-const FilmsSlider = memo<FilmSliderT>(function FilmsSlider({ films, isMobile }) {
-  useMobilePaddings(isMobile);
+export const FilmsSlider = memo<FilmSliderT>(function FilmsSlider({ films, isMobile }) {
   const matchKaraoke = useMatch({ path: 'karaoke' });
+
+  const [emblaRef, emblaApi] = useEmblaCarousel({
+    align: 'start',
+    containScroll: 'trimSnaps',
+    slidesToScroll: 1,
+    dragFree: true,
+  });
+
+  const scrollPrev = useCallback(() => emblaApi?.scrollPrev(), [emblaApi]);
+  const scrollNext = useCallback(() => emblaApi?.scrollNext(), [emblaApi]);
 
   if (matchKaraoke && isMobile) return null;
 
@@ -18,22 +26,31 @@ const FilmsSlider = memo<FilmSliderT>(function FilmsSlider({ films, isMobile }) 
     <div>
       <div className={s.cinemaSlider}>
         <h4 className={`displayMobile ${s.mobileTitle}`}>Фильмы</h4>
-        <Splide
-          options={{
-            perPage: isMobile ? 3 : 5,
-            perMove: 1,
-            pagination: false,
-            gap: isMobile ? '2vw' : '0.6rem',
-            arrows: !isMobile,
-          }}
-          className={s.splideWithArrows}
-        >
-          {films.map((item, i) => (
-            <Slide film={item} key={i + 'FS'} />
-          ))}
-        </Splide>
-      </div>
+        <div style={{ position: 'relative' }}>
+          <div className={s.viewport} ref={emblaRef}>
+            <div className={s.track}>
+              {films.map((film, i) => (
+                <Slide film={film} key={i + 'FS'} />
+              ))}
+            </div>
+          </div>
 
+          {!isMobile && (
+            <>
+              <button aria-label="Previous" className={s.arrowPrev} onClick={scrollPrev}>
+                <svg className={s.arrowSvg} viewBox="0 0 24 24">
+                  <path d="M15.41 7.41L14 6l-6 6 6 6 1.41-1.41L10.83 12z" />
+                </svg>
+              </button>
+              <button aria-label="Next" className={s.arrowNext} onClick={scrollNext}>
+                <svg className={s.arrowSvg} viewBox="0 0 24 24">
+                  <path d="M10 6L8.59 7.41 13.17 12l-4.58 4.59L10 18l6-6z" />
+                </svg>
+              </button>
+            </>
+          )}
+        </div>
+      </div>
       <div className="separatorMobile" />
     </div>
   );
@@ -41,14 +58,12 @@ const FilmsSlider = memo<FilmSliderT>(function FilmsSlider({ films, isMobile }) 
 
 const Slide = memo(function Slide({ film }: SlideT) {
   const title = useMemo(() => {
-    const baseTitle = film.pirate
-      ? `${film.title} ${PRE_SHOW_SERVICE}`
-      : film.title;
-    return removeLineBreaks(baseTitle);
+    const base = film.pirate ? `${film.title} ${PRE_SHOW_SERVICE}` : film.title;
+    return removeLineBreaks(base);
   }, [film.title, film.pirate]);
 
   return (
-    <SplideSlide className="swSlide">
+    <div className={`swSlide ${s.slide}`}>
       <Link className="swSlide__a" to={`movies/${film.link}`}>
         <FilmImg
           age={film.age}
@@ -60,7 +75,7 @@ const Slide = memo(function Slide({ film }: SlideT) {
         <h1 className={s.slideH1}>{title}</h1>
         <p className={s.slideP}>{film.beginDate}</p>
       </Link>
-    </SplideSlide>
+    </div>
   );
 });
 
@@ -68,8 +83,5 @@ type SlideT = {
   film: CinemaStateT['films'][0];
 };
 type FilmSliderT = {
-  films: CinemaStateT['films'];
-  isMobile: boolean;
+  films: CinemaStateT['films']; isMobile: boolean;
 };
-
-export default FilmsSlider;
